pyproject.toml
[project]
name = "django-react-boilerplate"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "yourname", email = "youremail@gmail.com"},
]
dependencies = [
    "django>=4.2.15",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
    "django-split-settings>=1.3.2",
]
requires-python = "==3.9.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
